refactor: move print and input functions to utils.py
